refactor: LD Provider in web with associated utils
