+++ libs/shared-web/src/providers/index.ts
@@ -1,3 +1,2 @@
 export * from './SegmentProvider';
 export * from './AuthProvider';
-export * from './LaunchDarklyProvider';
