+++ libs/shared-web/src/utils/auth-selectors/index.ts
@@ -1,3 +1,1 @@
 export * from './selectHasUserCompletedSignUp';
-export * from './selectShouldShowLaunchDarklyFallback';
-export * from './selectShouldInitializeLaunchDarkly';
